apps/pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "workradar-pipeline"
version = "0.1.0"
description = "AI-powered job ingestion pipeline for WorkRadar — scrapes, parses, embeds, and matches jobs."
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "WorkRadar Team" }]

dependencies = [
    "httpx>=0.27,<1",
    "beautifulsoup4>=4.12,<5",
    "playwright>=1.44,<2",
    "openai>=1.30,<2",
    "anthropic>=0.34,<1",
    "numpy>=1.26,<3",
    "supabase>=2.5,<3",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "click>=8.1,<9",
    "python-dotenv>=1.0,<2",
    "tenacity>=8.3,<9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
workradar-pipeline = "main:cli"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
